# Reimplement request attribute access w/ __getattr__