Supersedes #5576. Many thanks to @yotamofek for the original PR.