@carltongibson 2017-11-22
@rpkilby OK, too early... Tests pass but if `__getattribute__` raises the `AttributeError`, why is `__getattr__` not then (re-)called?