@rpkilby 2017-11-22
I might have this *completely* wrong, as I don't know much about the cpython internals, but... my understanding is that the [underlying c code](https://github.com/python/cpython/blob/8c663fd60ecba9c82aa4c404dbfb1aae69fe8553/Objects/typeobject.c#L6366-L6406) calls `__getattribute__` and if an `AttributeError` is raised, it then calls `__getattr__` (if defined). The default implementation of `__getattribute__` isn't directly calling `__getattr__`, which is why we don't see an infinite loop. 

In essence:
- Underlying C code calls `Request.__getattribute__`  which fails and raises an `AttributeError`.
- Underlying C code calls `Request.__getattr__`, which checks the WSGIRequest. 
- WSGIRequest check fails, so `Requst.__getattr__` calls `Request.__getattribute__` to raise the exception.