@rpkilby 2017-11-22
I just did a simple timeit test, and the PR does seem to be faster (which makes sense, since there are less caught exceptions. I used the following test:

```python
class Perf(TestCase):
    def test_perf(self):
        from timeit import repeat

        wsgi_request = factory.get('/')
        request = Request(wsgi_request)

        # non-existent attribute
        def foo():
            try:
                request.foo
            except AttributeError:
                pass

        # attribute exists on WSGIRequest
        def body():
            request.body

        # attribute exists on DRF Request
        def _request():
            request._request

        print(min(repeat(foo, number=100000)))
        print(min(repeat(body, number=100000)))
        print(min(repeat(_request, number=100000)))

        raise Exception('done')
```

Timings | foo | body | _request
--- | ---- | --- | --- 
before | 0.57 s | 0.36 s | 0.07 s
after | 0.25 s | 0.13 s | 0.01 s