@carltongibson 2017-11-22
OK. It does read that way. Good linking skills.